refactor: introduce new codec trait to support various encoding
