+++ src/mito2/benches/memtable_bench.rs
@@ -25,6 +25,7 @@ use mito2::memtable::partition_tree::{PartitionTreeConfig, PartitionTreeMemtable
 use mito2::memtable::time_series::TimeSeriesMemtable;
 use mito2::memtable::{KeyValues, Memtable};
 use mito2::region::options::MergeMode;
+use mito2::row_converter::DensePrimaryKeyCodec;
 use mito2::test_util::memtable_util::{self, region_metadata_to_row_schema};
 use rand::rngs::ThreadRng;
 use rand::seq::SliceRandom;
@@ -43,8 +44,14 @@ fn write_rows(c: &mut Criterion) {
     // Note that this test only generate one time series.
     let mut group = c.benchmark_group("write");
     group.bench_function("partition_tree", |b| {
-        let memtable =
-            PartitionTreeMemtable::new(1, metadata.clone(), None, &PartitionTreeConfig::default());
+        let codec = Arc::new(DensePrimaryKeyCodec::new(&metadata));
+        let memtable = PartitionTreeMemtable::new(
+            1,
+            codec,
+            metadata.clone(),
+            None,
+            &PartitionTreeConfig::default(),
+        );
         let kvs =
             memtable_util::build_key_values(&metadata, "hello".to_string(), 42, &timestamps, 1);
         b.iter(|| {
@@ -71,7 +78,8 @@ fn full_scan(c: &mut Criterion) {
     let mut group = c.benchmark_group("full_scan");
     group.sample_size(10);
     group.bench_function("partition_tree", |b| {
-        let memtable = PartitionTreeMemtable::new(1, metadata.clone(), None, &config);
+        let codec = Arc::new(DensePrimaryKeyCodec::new(&metadata));
+        let memtable = PartitionTreeMemtable::new(1, codec, metadata.clone(), None, &config);
         for kvs in generator.iter() {
             memtable.write(&kvs).unwrap();
         }
@@ -108,7 +116,8 @@ fn filter_1_host(c: &mut Criterion) {
     let mut group = c.benchmark_group("filter_1_host");
     group.sample_size(10);
     group.bench_function("partition_tree", |b| {
-        let memtable = PartitionTreeMemtable::new(1, metadata.clone(), None, &config);
+        let codec = Arc::new(DensePrimaryKeyCodec::new(&metadata));
+        let memtable = PartitionTreeMemtable::new(1, codec, metadata.clone(), None, &config);
         for kvs in generator.iter() {
             memtable.write(&kvs).unwrap();
         }
