+++ src/mito2/src/test_util/memtable_util.rs
@@ -36,7 +36,9 @@ use crate::memtable::{
     BoxedBatchIterator, BulkPart, KeyValues, Memtable, MemtableBuilder, MemtableId, MemtableRanges,
     MemtableRef, MemtableStats,
 };
-use crate::row_converter::{McmpRowCodec, RowCodec, SortField};
+use crate::row_converter::{
+    DensePrimaryKeyCodec, PrimaryKeyCodec, PrimaryKeyCodecExt, PrimaryKeyEncoder, SortField,
+};
 
 /// Empty memtable for test.
 #[derive(Debug, Default)]
@@ -313,12 +315,7 @@ pub(crate) fn encode_keys(
     key_values: &KeyValues,
     keys: &mut Vec<Vec<u8>>,
 ) {
-    let row_codec = McmpRowCodec::new(
-        metadata
-            .primary_key_columns()
-            .map(|c| SortField::new(c.column_schema.data_type.clone()))
-            .collect(),
-    );
+    let row_codec = DensePrimaryKeyCodec::new(metadata);
     for kv in key_values.iter() {
         let key = row_codec.encode(kv.primary_keys()).unwrap();
         keys.push(key);
@@ -327,7 +324,7 @@ pub(crate) fn encode_keys(
 
 /// Encode one key.
 pub(crate) fn encode_key_by_kv(key_value: &KeyValue) -> Vec<u8> {
-    let row_codec = McmpRowCodec::new(vec![
+    let row_codec = DensePrimaryKeyCodec::with_fields(vec![
         SortField::new(ConcreteDataType::string_datatype()),
         SortField::new(ConcreteDataType::uint32_datatype()),
     ]);
