+++ src/store-api/src/metadata.rs
@@ -523,7 +523,7 @@ impl RegionMetadataBuilder {
             column_metadatas: vec![],
             primary_key: vec![],
             schema_version: 0,
-            primary_key_encoding: PrimaryKeyEncoding::Full,
+            primary_key_encoding: PrimaryKeyEncoding::Dense,
         }
     }
 
@@ -1538,7 +1538,7 @@ mod test {
     fn test_region_metadata_deserialize_default_primary_key_encoding() {
         let serialize = r#"{"column_metadatas":[{"column_schema":{"name":"a","data_type":{"Int64":{}},"is_nullable":false,"is_time_index":false,"default_constraint":null,"metadata":{}},"semantic_type":"Tag","column_id":1},{"column_schema":{"name":"b","data_type":{"Float64":{}},"is_nullable":false,"is_time_index":false,"default_constraint":null,"metadata":{}},"semantic_type":"Field","column_id":2},{"column_schema":{"name":"c","data_type":{"Timestamp":{"Millisecond":null}},"is_nullable":false,"is_time_index":false,"default_constraint":null,"metadata":{}},"semantic_type":"Timestamp","column_id":3}],"primary_key":[1],"region_id":5299989648942,"schema_version":0}"#;
         let deserialized: RegionMetadata = serde_json::from_str(serialize).unwrap();
-        assert_eq!(deserialized.primary_key_encoding, PrimaryKeyEncoding::Full);
+        assert_eq!(deserialized.primary_key_encoding, PrimaryKeyEncoding::Dense);
 
         let serialize = r#"{"column_metadatas":[{"column_schema":{"name":"a","data_type":{"Int64":{}},"is_nullable":false,"is_time_index":false,"default_constraint":null,"metadata":{}},"semantic_type":"Tag","column_id":1},{"column_schema":{"name":"b","data_type":{"Float64":{}},"is_nullable":false,"is_time_index":false,"default_constraint":null,"metadata":{}},"semantic_type":"Field","column_id":2},{"column_schema":{"name":"c","data_type":{"Timestamp":{"Millisecond":null}},"is_nullable":false,"is_time_index":false,"default_constraint":null,"metadata":{}},"semantic_type":"Timestamp","column_id":3}],"primary_key":[1],"region_id":5299989648942,"schema_version":0,"primary_key_encoding":"sparse"}"#;
         let deserialized: RegionMetadata = serde_json::from_str(serialize).unwrap();
